% Generated by build script
\name{off_compare}
\alias{off_compare}
\title{Compare nutrition between multiple products}
\usage{
off_compare()
}
\arguments{
\item{barcodes}{Character vector of product barcodes}
}
\value{tibble: code, product_name, energy_kcal, fat_100g, carbs_100g,}
\description{Compare nutrition between multiple products}
