% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/off.R
\name{off_context}
\alias{off_context}
\title{Show Open Food Facts package context for LLM use}
\usage{
off_context()
}
\value{
Invisible string with full context
}
\description{
Show Open Food Facts package context for LLM use
}
