% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/off.R
\name{off_product}
\alias{off_product}
\title{Get a specific product by barcode}
\usage{
off_product(barcode)
}
\arguments{
\item{barcode}{Product barcode (EAN-13, UPC, etc.)}
}
\value{
tibble with one row: code, product_name, brands, categories,
ingredients_text, nutriscore_grade, energy_kcal, fat, carbohydrates,
proteins, image_url
}
\description{
Get a specific product by barcode
}
